requirements.txt
appdirs==1.4.4
asgiref==3.2.10
attrs==19.3.0
Babel==2.8.0
black==19.10b0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
defusedxml==0.6.0
dj-rest-auth==1.1.0
Django==3.0.8
django-allauth==0.42.0
django-countries==6.1.2
django-filter==2.3.0
django-modeltranslation==0.15.1
django-phonenumber-field==4.0.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
docutils==0.16
flake8==3.8.3
idna==2.10
importlib-metadata==1.7.0
mccabe==0.6.1
oauthlib==3.1.0
pathspec==0.8.0
phonenumbers==8.12.7
Pillow==7.2.0
psycopg2-binary==2.8.5
pycodestyle==2.6.0
pyflakes==2.2.0
PyJWT==1.7.1
python-dotenv==0.14.0
python3-openid==3.2.0
pytz==2020.1
regex==2020.7.14
requests==2.24.0
requests-oauthlib==1.3.0
rstcheck==3.3.1
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
zipp==3.1.0
